refactor: split plugin function into parts
